Fixed bug in maximum_normed_residual
Previously, when a data set was passed to
`maximum_normed_residual` where all the
observations but two would be considered
outliers, the critical value would be calculated
with all but two observations removed. In that
loop, qt() would be called with df=0 since
the degrees of freedom are equal to the number
of remaining observations minus 2. qt() fails
when df=0, so the maximum_normed_residual
function would fail.